drm/amdkfd: kill user processes after hotplug out
At hotplug-out user processes seems to spin until a sigkill is
sent. In this version we tried to let amdkfd to initiate the kill.
This solution is not suited for upstream drm because a user
application that had kfd open and is not accessing the GPU device.  In
this case, kfd sending SIGBUS might kill the user process unneccarily.
In hotplug-out scenario, we should let user to terminate the process.

The following are the techniques that kill user processes at PCIe hot removal.

1. Kill process synchronously by sending SIGBUS to all the user
processes, before actual work of pcie removal is done.

2. Sleep for 5s to let sdma ip block to drain during unplanned
hotplug out.

Signed-off-by: Shuotao Xu <shuotaoxu@microsoft.com>